// src/generated-twilio-api-models/twilio_messaging_v1.schema.d.ts
export interface components {
  schemas: {
// ...
    "messaging.v1.brand_registrations": {
      /** @description The unique string to identify Brand Registration. */
      sid?: string | null;
      /** @description The SID of the [Account](https://www.twilio.com/docs/iam/api/account) that created the Brand Registration resource. */
      account_sid?: string | null;
      /** @description A2P Messaging Profile Bundle BundleSid. */
      customer_profile_bundle_sid?: string | null;
      /** @description A2P Messaging Profile Bundle BundleSid. */
      a2p_profile_bundle_sid?: string | null;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was created specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_created?: string | null;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was last updated specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_updated?: string | null;
      /** @description Type of brand. One of: "STANDARD", "SOLE_PROPRIETOR". SOLE_PROPRIETOR is for the low volume, SOLE_PROPRIETOR campaign use case. There can only be one SOLE_PROPRIETOR campaign created per SOLE_PROPRIETOR brand. STANDARD is for all other campaign use cases. Multiple campaign use cases can be created per STANDARD brand. */
      brand_type?: string | null;
      /** @description Brand Registration status. One of "PENDING", "APPROVED", "FAILED", "IN_REVIEW", "DELETED". */
      status?: components["schemas"]["brand_registrations_enum_status"];
      /** @description Campaign Registry (TCR) Brand ID. Assigned only after successful brand registration. */
      tcr_id?: string | null;
      /** @description DEPRECATED. A reason why brand registration has failed. Only applicable when status is FAILED. */
      failure_reason?: string | null;
      /** @description A list of errors that occurred during the brand registration process. */
      errors?: unknown[] | null;
      /**
       * Format: uri
       * @description The absolute URL of the Brand Registration resource.
       */
      url?: string | null;
      /** @description The secondary vetting score if it was done. Otherwise, it will be the brand score if it's returned from TCR. It may be null if no score is available. */
      brand_score?: number | null;
      /** @description DEPRECATED. Feedback on how to improve brand score */
      brand_feedback?: components["schemas"]["brand_registrations_enum_brand_feedback"][] | null;
      /** @description When a brand is registered, TCR will attempt to verify the identity of the brand based on the supplied information. */
      identity_status?: components["schemas"]["brand_registrations_enum_identity_status"];
      /** @description Publicly traded company identified in the Russell 3000 Index */
      russell_3000?: boolean | null;
      /** @description Identified as a government entity */
      government_entity?: boolean | null;
      /** @description Nonprofit organization tax-exempt status per section 501 of the U.S. tax code. */
      tax_exempt_status?: string | null;
      /** @description A flag to disable automatic secondary vetting for brands which it would otherwise be done. */
      skip_automatic_sec_vet?: boolean | null;
      /** @description A boolean that specifies whether brand should be a mock or not. If true, brand will be registered as a mock brand. Defaults to false if no value is provided. */
      mock?: boolean | null;
      /** Format: uri-map */
      links?: Record<string, unknown> | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {
// ...
  ListPhoneNumber: {
    parameters: {
      query?: {
        /** @description How many resources to return in each list page. The default is 50, and the maximum is 1000. */
        PageSize?: number;
        /** @description The page index. This value is simply for client state. */
        Page?: number;
        /** @description The page token. This is provided by the API. */
        PageToken?: string;
      };
      path: {
        /** @description The SID of the [Service](https://www.twilio.com/docs/chat/rest/service-resource) to read the resources from. */
        ServiceSid: string;
      };
    };
    responses: {
      /** @description OK */
      200: {
        content: {
          "application/json": {
            phone_numbers?: components["schemas"]["messaging.v1.service.phone_number"][];
            meta?: {
              /** Format: uri */
              first_page_url?: string;
              /** Format: uri */
              next_page_url?: string | null;
              page?: number;
              page_size?: number;
              /** Format: uri */
              previous_page_url?: string | null;
              /** Format: uri */
              url?: string;
              key?: string;
            };
          };
        };
      };
    };
  };
// ...
  ListUsAppToPerson: {
    parameters: {
      query?: {
        /** @description How many resources to return in each list page. The default is 50, and the maximum is 1000. */
        PageSize?: number;
        /** @description The page index. This value is simply for client state. */
        Page?: number;
        /** @description The page token. This is provided by the API. */
        PageToken?: string;
      };
      path: {
        /** @description The SID of the [Messaging Service](https://www.twilio.com/docs/messaging/api/service-resource) to fetch the resource from. */
        MessagingServiceSid: string;
      };
    };
    responses: {
      /** @description OK */
      200: {
        content: {
          "application/json": {
            compliance?: components["schemas"]["messaging.v1.service.us_app_to_person"][];
            meta?: {
              /** Format: uri */
              first_page_url?: string;
              /** Format: uri */
              next_page_url?: string | null;
              page?: number;
              page_size?: number;
              /** Format: uri */
              previous_page_url?: string | null;
              /** Format: uri */
              url?: string;
              key?: string;
            };
          };
        };
      };
    };
  };
// ...
}