// src/generated-twilio-api-models/twilio_api_v2010.schema.d.ts
export interface components {
  schemas: {
// ...
    "api.v2010.account.address": {
      /** @description The SID of the [Account](https://www.twilio.com/docs/iam/api/account) that is responsible for the Address resource. */
      account_sid?: string | null;
      /** @description The city in which the address is located. */
      city?: string | null;
      /** @description The name associated with the address.This property has a maximum length of 16 4-byte characters, or 21 3-byte characters. */
      customer_name?: string | null;
      /**
       * Format: date-time-rfc-2822
       * @description The date and time in GMT that the resource was created specified in [RFC 2822](https://www.ietf.org/rfc/rfc2822.txt) format.
       */
      date_created?: string | null;
      /**
       * Format: date-time-rfc-2822
       * @description The date and time in GMT that the resource was last updated specified in [RFC 2822](https://www.ietf.org/rfc/rfc2822.txt) format.
       */
      date_updated?: string | null;
      /** @description The string that you assigned to describe the resource. */
      friendly_name?: string | null;
      /**
       * Format: iso-country-code
       * @description The ISO country code of the address.
       */
      iso_country?: string | null;
      /** @description The postal code of the address. */
      postal_code?: string | null;
      /** @description The state or region of the address. */
      region?: string | null;
      /** @description The unique string that that we created to identify the Address resource. */
      sid?: string | null;
      /** @description The number and street address of the address. */
      street?: string | null;
      /** @description The URI of the resource, relative to `https://api.twilio.com`. */
      uri?: string | null;
      /** @description Whether emergency calling has been enabled on this number. */
      emergency_enabled?: boolean | null;
      /** @description Whether the address has been validated to comply with local regulation. In countries that require valid addresses, an invalid address will not be accepted. `true` indicates the Address has been validated. `false` indicate the country doesn't require validation or the Address is not valid. */
      validated?: boolean | null;
      /** @description Whether the address has been verified to comply with regulation. In countries that require valid addresses, an invalid address will not be accepted. `true` indicates the Address has been verified. `false` indicate the country doesn't require verified or the Address is not valid. */
      verified?: boolean | null;
      /** @description The additional number and street address of the address. */
      street_secondary?: string | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}